making sure "binsize" is saved as a double for NLMS compatibility
git-svn-id: http://citadel.physiology.northwestern.edu/svn/s1_analysis@303 984b1e3c-1c57-0410-ae62-c18f90d3b645

+++ BMI_analysis/BuildModel.m
@@ -20,7 +20,7 @@ function [filter, varargout]=BuildModel(binnedData, dataPath, fillen, UseAllInpu
         binnedData = LoadDataStruct(binnedData, 'binned');
     end
 
-    binsize = binnedData.timeframe(2)-binnedData.timeframe(1);
+    binsize = double(binnedData.timeframe(2)-binnedData.timeframe(1));
     
     if nargout > 2
         disp('Wrong number of output arguments');
